allow some deprecation warnings to be silenced
Previously, a deprecation warning was issued instructing the user to
load the core extensions explicitly.  This warning continued to be
issued even if the user added code to require the core extensions (as
instructed).  This change sets a global variable when the core
extensions are loaded appropriately so that the user isn't annoyed by
unnecessary deprecation warnings for the functionality they're no longer
making use of.
