+++ lib/money.rb
@@ -3,5 +3,5 @@ require "bigdecimal/util"
 require "i18n" rescue LoadError
 require "money/currency"
 require "money/money"
-require "money/core_extensions"
+require "money/core_extensions/core_extensions"
 require "money/deprecations"
